backend/requirements.txt
ultralytics
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
opencv-python
jupyter
ipykernel
